toolkit/content/tests/chrome/window_popup_preventdefault_chrome.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>

<window title="Popup Prevent Default Tests"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <title>Popup Prevent Default Tests</title>

<!--
  This tests checks that preventDefault can be called on a popupshowing
  event or popuphiding event to prevent the default behaviour.
  -->

<script>

var gBlockShowing = true;
var gBlockHiding = true;
var gShownNotAllowed = true;
var gHiddenNotAllowed = true;

var fm = Components.classes["@mozilla.org/focus-manager;1"].
           getService(Components.interfaces.nsIFocusManager);

var is = function(l, r, v) { window.opener.wrappedJSObject.SimpleTest.is(l, r, v); }
var isnot = function(l, r, v) { window.opener.wrappedJSObject.SimpleTest.isnot(l, r, v); }

function runTest()
{
  var menu = document.getElementById("menu");

  is(fm.activeWindow, window, "active window at start");
  is(fm.focusedWindow, window, "focused window at start");

  is(window.windowState, window.STATE_NORMAL, "window is normal");
  // the minimizing test sometimes fails on Linux so don't test it there
  if (navigator.platform.indexOf("Lin") == 0) {
    menu.open = true;
    return;
  }
  window.minimize();
  is(window.windowState, window.STATE_MINIMIZED, "window is minimized");

  isnot(fm.activeWindow, window, "active window after minimize");
  isnot(fm.focusedWindow, window, "focused window after minimize");

  menu.open = true;

  setTimeout(runTestAfterMinimize, 0);
}

function runTestAfterMinimize()
{
  var menu = document.getElementById("menu");
  is(menu.firstChild.state, "closed", "popup not opened when window minimized");

  window.restore();
  is(window.windowState, window.STATE_NORMAL, "window is restored");

  is(fm.activeWindow, window, "active window after restore");
  is(fm.focusedWindow, window, "focused window after restore");

  menu.open = true;
}

function popupShowing(event)
{
  if (gBlockShowing) {
    event.preventDefault();
    gBlockShowing = false;
    setTimeout(function() {
                 gShownNotAllowed = false;
                 document.getElementById("menu").open = true;
               }, 3000, true);
  }
}

function popupShown()
{
  window.opener.wrappedJSObject.SimpleTest.ok(!gShownNotAllowed, "popupshowing preventDefault");
  document.getElementById("menu").open = false;
}

function popupHiding(event)
{
  if (gBlockHiding) {
    event.preventDefault();
    gBlockHiding = false;
    setTimeout(function() {
                 gHiddenNotAllowed = false;
                 document.getElementById("menu").open = false;
               }, 3000, true);
  }
}

function popupHidden()
{
  window.opener.wrappedJSObject.SimpleTest.ok(!gHiddenNotAllowed, "popuphiding preventDefault");
  window.opener.wrappedJSObject.SimpleTest.finish();
  window.close();
}

window.opener.wrappedJSObject.SimpleTest.waitForFocus(runTest, window);
</script>

<button id="menu" type="menu" label="Menu">
  <menupopup onpopupshowing="popupShowing(event);"
             onpopupshown="popupShown();"
             onpopuphiding="popupHiding(event);"
             onpopuphidden="popupHidden();">
    <menuitem label="Item"/>
  </menupopup>
</button>


</window>
